refactor : 패키지 경로 및 JavaDoc 추가
