+++ src/main/java/com/nhnacademy/frontservice/dto/token/TokenRequest.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontservice.dto;
+package com.nhnacademy.frontservice.dto.token;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
